fix: Fix type of factor in planner computations

+++ src/query/plan/cost_estimator.hpp
@@ -728,7 +728,7 @@ class CostEstimator : public HierarchicalLogicalOperatorVisitor {
     auto lower = BoundToPropertyValue(lower_bound);
     auto upper = BoundToPropertyValue(upper_bound);
 
-    int64_t factor = 1;
+    double factor = 1;
     if (upper || lower) {
       factor = db_accessor_->VerticesCount(property, lower, upper);
     } else {
